+++ pypeit/core/skysub.py
@@ -619,7 +619,7 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, rn2_img, t
     if std is True:
         chi2_sigrej = 100.0
         #sigrej_ceil = 1e10
-        sigrej = 25.0
+        sigrej = 50.0  # 25 wasn't enough for MagE 2x2 binning (probably undersampled)
     else:
         chi2_sigrej = 6.0
         #sigrej_ceil = 10.0