+++ pypeit/reduce.py
@@ -596,6 +596,7 @@ class MultiSlit(Reduce):
                                 sig_thresh=sig_thresh, hand_extract_dict=manual_extract_dict,
                                 specobj_dict=specobj_dict, show_peaks=show_peaks,
                                 show_fits=show_fits, show_trace=show_trace,
+                                trim_edg=self.redux_par['find_trim_edge'],
                                 qa_title=qa_title, nperslit=self.redux_par['maxnumber'])
             sobjs.add_sobj(sobjs_slit)
 
@@ -725,6 +726,7 @@ class Echelle(Reduce):
                                 plate_scale=plate_scale, std_trace=std_trace,
                                 specobj_dict=specobj_dict,sig_thresh=sig_thresh,
                                 show_peaks=show_peaks, show_fits=show_fits,
+                                trim_edg=self.redux_par['find_trim_edge'],
                                 show_trace=show_trace, debug=debug)
 
         # Steps