+++ pypeit/scripts/coadd_1dspec.py
@@ -39,12 +39,16 @@ def main(args, unit_test=False, path=''):
     from pypeit.core import coadd1d
     from pypeit.core import load
     from pypeit import specobjs
+    from pypeit.spectrographs import util
 
     from IPython import embed
 
     # Load the input file
     with open(args.infile, 'r') as infile:
         coadd_dict = yaml.load(infile)
+
+    # Spectrograph
+    spectrograph = util.load_spectrograph(coadd_dict.pop('spectrograph'))
 
     # Grab object names in the spectra
     filelist = coadd_dict.pop('filenames')