+++ CHANGES.rst
@@ -13,6 +13,12 @@
 - Started notes on how to generate a new spectrograph in PypeIt
 - Refactoring of reduce to take a ScienceImage object for the images and the mask
 - Updates to many spectrograph files to put datasec, oscansec in the raw frame
+- Add find_trim_edge and std_prof_nsigma parameters
+- A bit of tuning for MagE
+- Fixes for Echelle in fluxspec
+- Writes a chosen set of header cards to the spec1D and coadd files
+- Updates for FORS2
+
 
 0.10.1 (22 May 2019)
 --------------------